Prettier outputs for benchmarks

+++ quiet-gar/src/test.c
@@ -175,7 +175,7 @@ int main(int argc, char **argv)
     printf("******************************************\n");
     printf("SUMMARY\n");
     printf("Avg. time: %.2f ms to generate %llu OTs\n", avg, (long long unsigned int)NUM_OTS);
-    printf("Performance: %.2f OTs/sec\n", ((double)(NUM_OTS) / avg) * 1000); // Convert ms to seconds
+    printf("Performance: %i OTs/sec\n", (int)(((double)(NUM_OTS) / avg) * 1000)); // Convert ms to seconds
     printf("Number of trials: %i\n", testTrials);
     printf("******************************************\n\n");
 }
